Refactor PPO model creation: move make_actor and make_critic to models.py
